Overhaul build system with BUILD_CONTEXT resolver and error_in_r mode
- Implement BUILD_CONTEXT resolver: dev-monorepo, dev-detached, vendored, prepare-cran
- Add #[miniextendr(error_in_r)] mode for R-level error handling
- Use cargo package for vendoring workspace crates
- Track Cargo.toml and document.rs directly in git
- Replace panic-as-control-flow with r_stop in generated wrappers
- Fix env-class trait method dispatch and invisible returns
- Eliminate miniextendr R package runtime dependency for trait ABI

+++ minirextendr/inst/templates/monorepo/rpkg/lib.rs
@@ -1,12 +1,12 @@
 use miniextendr_api::{miniextendr, miniextendr_module};
 
-/// A simple function that adds two integers
+/// A simple function that adds two numbers
 ///
-/// @param a First integer
-/// @param b Second integer
+/// @param a First number
+/// @param b Second number
 /// @return Sum of a and b
 #[miniextendr]
-pub fn add(a: i32, b: i32) -> i32 {
+pub fn add(a: f64, b: f64) -> f64 {
     a + b
 }
